+++ src/components/Helpers/CustomRenderCell.tsx
@@ -8,7 +8,7 @@ export default function CustomRenderCell({ value, isCopyToClipboard = false }) {
     //     {value || <NoDataCell />}
     //     { inHover && isCopyToClipboard && value ? <CopyToClipboard textToCopy={value} /> : null}
     // </span>
-    return <> {value ? <span title={value} className="text-truncate">{value}</span> : <NoDataCell />}
+    return <> {value ? <span title={value} className="text-truncate">{value} 123</span> : <NoDataCell />}
         {isCopyToClipboard && value ? <CopyToClipboard textToCopy={value} /> : null}
     </>
 }