+++ src/pages/Opportunities/index.tsx
@@ -433,6 +433,8 @@ const Opportunities = () => {
                   afterImportCompleted={() => {
                     fetchOpportunities();
                   }}
+                  isExportAllOrSomeFeature={true}
+                  total={rowCount}
                   recordsToExport={selectedRecords.length}
                   ids={selectedRecords.length ? selectedRecords.map((obj) => obj._id) : []}
                   onExportToExcelSuccess={() => {
@@ -490,6 +492,7 @@ const Opportunities = () => {
           actionWidth={100}
           loading={loading}
           renderedFrom={opportunityResource}
+          refreshGrid={fetchOpportunities}
         />
 
         {showDeleteWarningConfirmBox ? (