+++ src/pages/Product/index.tsx
@@ -3,10 +3,9 @@ import Grid from '@material-ui/core/Grid';
 import Button from '@material-ui/core/Button';
 import CustomBreadCrumbs from "../../components/CustomBreadCrumbs";
 import AddIcon from "@material-ui/icons/Add";
-import Tooltip from "@material-ui/core/Tooltip";
 import IconButton from '@material-ui/core/IconButton';
 import DeleteIcon from '@material-ui/icons/Delete';
-import { Link } from 'react-router-dom'
+import { Link, useHistory } from 'react-router-dom'
 import { CustomToastContext } from "../../StateProvider/CustomToastContext/CustomToastContext";
 import axiosInstance from "../../axios/axiosInstance";
 import CreateProduct from "../../components/Product/CreateProduct";
@@ -30,6 +29,9 @@ import CommonSkeleton from "../../components/Helpers/CommonSkeleton";
 import NoDataCell from "../../components/Helpers/NoDataCell";
 import { useData } from "../../StateProvider/Provider";
 import { sortBy } from 'lodash';
+import HtmlTooltip from '../../components/CustomTooltipTitle'
+import { RiBillLine } from "react-icons/ri";
+
 
 const ignoreField = ["qty", "priceTemplate"]
 
@@ -44,6 +46,7 @@ var levalOrderBy = [
 
 const Product = () => {
 
+    const history = useHistory();
     const toastConfig = useContext(CustomToastContext)
     const [open, setOpen] = useState(false);
     const [productId, setProductId] = useState(null);
@@ -315,7 +318,7 @@ const Product = () => {
     const ActionsRenderer = params => (
         <>
             {productPermissions.isCreate &&
-                <Tooltip title="Clone">
+                <HtmlTooltip title="Clone">
                     <IconButton
                         size="small"
                         aria-label="Clone"
@@ -323,17 +326,26 @@ const Product = () => {
                     >
                         <FileCopyIcon color="primary" />
                     </IconButton>
-                </Tooltip>
+                </HtmlTooltip>
             }
             {productPermissions.isDelete &&
-                <Tooltip title="Delete">
+                <HtmlTooltip title="Delete">
                     <IconButton size="small" aria-label="Delete" onClick={() => {
                         setDeleteRecord(params.data);
                         setShowDeleteConfirmBox(true)
                     }} >
                         <DeleteIcon color="error" />
                     </IconButton>
-                </Tooltip >
+                </HtmlTooltip >
+            }
+            {productPermissions.isRead &&
+                <HtmlTooltip title="BOM">
+                    <IconButton size="small" aria-label="View BOM" onClick={() => {
+                        history.push(`${routes.productDetail.path}/${params.data._id}/bom`, { productName: params.data.productName })
+                    }} >
+                        <RiBillLine color="primary" />
+                    </IconButton>
+                </HtmlTooltip >
             }
         </>
     )
@@ -399,6 +411,14 @@ const Product = () => {
                         if (isImportedSuccessfully) {
                             fetchProduct();
                         }
+                    }}
+                    isExportAllOrSomeFeature={true}
+                    total={rowCount}
+                    recordsToExport={selectedRecords.length}
+                    ids={selectedRecords.length ? selectedRecords.map((obj) => obj._id) : []}
+                    onExportToExcelSuccess={() => {
+                        if (gridApi) gridApi.deselectAll()
+                        else fetchProduct()
                     }}
                 />
             </Grid>
@@ -466,6 +486,7 @@ const Product = () => {
                     actionWidth={150}
                     loading={loading}
                     renderedFrom="productPage"
+                    refreshGrid={fetchProduct}
                 />
                 : <Box p={2} height={500} bgcolor="white"><CommonSkeleton lenArray={[...Array(10).keys()]} /></Box>}
         </div>
@@ -474,6 +495,7 @@ const Product = () => {
                 isClone={isClone}
                 productId={productId}
                 handleClose={handleClose}
+                isRedirectToDetailPage={true}
                 openFrom="productMaster"
             />
         }