+++ src/pages/ProductCategory/index.tsx
@@ -373,8 +373,17 @@ const ProductCategory = () => {
                     afterImportCompleted={() => {
                         fetchProductCategory();
                     }}
+                    isExportAllOrSomeFeature={true}
+                    total={rowCount}
+                    recordsToExport={selectedRecords.length}
+                    ids={selectedRecords.length ? selectedRecords.map((obj) => obj._id) : []}
+                    onExportToExcelSuccess={() => {
+                        if (gridApi) gridApi.deselectAll()
+                        else fetchProductCategory()
+                    }}
                 />
             </Grid>
+
         </Grid>
         <CustomContainer>
             <div className="header-panel">
@@ -437,7 +446,9 @@ const ProductCategory = () => {
                 setGridApi={setGridApi}
                 dispatch={dispatch}
                 rowCount={rowCount} limit={limit} pageSizes={pageSizes} page={page} allowAction={true}
-                loading={loading} renderedFrom="productCategoryPage" />
+                loading={loading} renderedFrom="productCategoryPage"
+                refreshGrid={fetchProductCategory}
+            />
 
             {showDeleteConfirmBox &&
                 <ConfirmationDialog